🛡️ Sentinel: [HIGH] Fix XSS by removing dangerouslySetInnerHTML
- Converted `renderJsonWithHighlighting` to return an array of React elements instead of a concatenated HTML string.
- Removed manual HTML escaping since React handles escaping automatically for text nodes.
- Updated `ToolResultModal.tsx` and `jsonHighlighter.tsx` to stop using `dangerouslySetInnerHTML`.
- Fixed CI test failures by mocking React environment properly in `useInactivity` test and explicitly mocking encryption keys in `Encryption` tests.
- Appended critical learning entry to `.jules/sentinel.md`.
